+++ bot/utility/configs.py
@@ -43,9 +43,9 @@ EVENT_ANNOUNCEMENT_CONFIG = {
         "reactions": "<:RegimentalColours:863525506779054120> if you are attending the 63e event\n❓ if you are unsure of your attendance\n❌ if you cannot attend today\n<:KingsColours:863525505515782184> if you are attending the Public Linebattle\n❔ if you are unsure of your attendance\n❎ if you cannot attend today"
     },
     "Wednesday": {
-        "numEvents": "two events",
-        "eventBody": "- 19:00 GMT / 14:00 EST: 51st Wednesday Linebattle\n- 20:00 GMT / 15:00 EST: Wednesday EU8 Public Linebattle",
-        "reactions": "<:RegimentalColours:863525506779054120> if you are attending the 51st event\n❓ if you are unsure of your attendance\n❌ if you cannot attend today\n<:KingsColours:863525505515782184> if you are attending the Public Linebattle\n❔ if you are unsure of your attendance\n❎ if you cannot attend today"
+        "numEvents": "one event",
+        "eventBody": "- 19:00 GMT / 14:00 EST: 51st Wednesday Linebattle",
+        "reactions": "<:RegimentalColours:863525506779054120> if you are attending the 51st event\n❓ if you are unsure of your attendance\n❌ if you cannot attend today"
     },
     "Thursday": {
         "numEvents": 0